Enhanced app functionality and structure
- Added DbSet properties for entities in MatLidStoreDatabaseContext.cs to represent tables in the database.
- Introduced IsProductUnique method in IProductRepository for product uniqueness check.
- Modified OrderDto and OrderDetailDto with additional properties for enhanced data transfer.
- Implemented CQRS pattern with commands and queries for orders and products management.
- Removed CreateAddressCommand and handler, indicating a shift in address management.
- Added FluentValidation for CreateOrderCommandValidator, CreateProductCommandValidator, and UpdateProductCommandValidator to ensure data integrity.
- Enhanced exception handling in command handlers with custom exceptions for better error reporting.
- Added AutoMapper profiles for efficient domain entities and DTOs mapping.
- Adjusted project structure in MLS.Application.csproj for better organization and feature management.

+++ src/MSL.Application/Features/Product/Commands/DeleteProductCommand/DeleteProductCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using MLS.Application.Contracts.Persistence.IRepositories;
+using MLS.Application.Exceptions;
 
 namespace MLS.Application.Features.Product.Commands.DeleteProductCommand
 {
@@ -18,6 +19,8 @@ namespace MLS.Application.Features.Product.Commands.DeleteProductCommand
             var productToDelete = await _productRepository.GetByIdAsync(request.Id);
 
             // Verify that record exists
+            if (productToDelete is null)
+                throw new NotFoundException(nameof(Domain.Entities.Product), request.Id);
 
             // Remove in database
             await _productRepository.DeleteAsync(productToDelete);
